@@ -4,23 +4,33 @@ declare (strict_types=1);
 
 namespace App\Http\Controllers;
 
-use App\Author;
-use App\Book;
+use App\DTOs\BookDto;
+use App\DTOs\BookReviewDto;
 use App\Filters\SearchAuthors;
 use App\Filters\SearchTitle;
 use App\Filters\Sort;
+use App\Http\Requests\BookListRequest;
 use App\Http\Requests\PostBookRequest;
 use App\Http\Requests\PostBookReviewRequest;
 use App\Http\Resources\BookResource;
 use App\Http\Resources\BookReviewResource;
-use Illuminate\Http\Request;
+use App\Models\Book;
+use App\Services\BookService\BookService;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Pipeline\Pipeline;
 use Illuminate\Support\Facades\Auth;
 
 class BooksController extends Controller
 {
-    public function getCollection(Request $request): AnonymousResourceCollection
+    private BookService $bookService;
+
+    public function __construct(BookService $bookService)
+    {
+        $this->bookService = $bookService;
+    }
+
+    public function index(BookListRequest $request): AnonymousResourceCollection
     {
         $books = app(Pipeline::class)
             ->send(Book::query())
@@ -35,31 +45,26 @@ class BooksController extends Controller
         return BookResource::collection($books);
     }
 
-    public function post(PostBookRequest $request): BookResource
+    public function store(PostBookRequest $request): BookResource
     {
-        $bookData = [
-            'isbn'        => $request->isbn,
-            'title'       => $request->title,
-            'description' => $request->description
-        ];
+        $bookDto = (new BookDto())
+            ->setTitle($request->title)
+            ->setIsbn($request->isbn)
+            ->setDescription($request->description)
+            ->setAuthorIds($request->authors);
 
-        /** @var Book $book */
-        $book = Book::query()->create($bookData);
-        $authors = Author::query()->findMany($request->authors);
-        $book->authors()->sync($authors);
-
+        $book = $this->bookService->store($bookDto);
         return BookResource::make($book);
     }
 
-    public function postReview(Book $book, PostBookReviewRequest $request): BookReviewResource
+    public function storeReview(Book $book, PostBookReviewRequest $request): JsonResponse
     {
-        $reviewData = [
-            'user_id' => Auth::id(),
-            'review'  => $request->review,
-            'comment' => $request->comment
-        ];
-        $review = $book->reviews()->create($reviewData);
+        $bookReviewDto = (new BookReviewDto())
+            ->setUserId(Auth::id())
+            ->setReview($request->get('review'))
+            ->setComment($request->get('comment'));
 
-        return BookReviewResource::make($review);
+        $review = $this->bookService->storeBookReview($book->id, $bookReviewDto);
+        return BookReviewResource::make($review)->response()->setStatusCode(201);
     }
 }
