@@ -1,24 +1,20 @@
 <?php
 
-/* @var $factory \Illuminate\Database\Eloquent\Factory */
+namespace Database\Factories;
 
-use App\Author;
-use Faker\Generator as Faker;
+use App\Models\Author;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+/**
+ * @extends Factory<Author>
+ */
+class  AuthorFactory extends Factory{
 
-$factory->define(Author::class, function (Faker $faker) {
-    return [
-        'name' => $faker->name,
-        'surname' => $faker->lastName,
-    ];
-});
+    public function definition(): array
+    {
+        return [
+            'name' => $this->faker->name,
+            'surname' => $this->faker->lastName,
+        ];
+    }
+}
