@@ -1,25 +1,22 @@
 <?php
 
-/* @var $factory \Illuminate\Database\Eloquent\Factory */
+namespace Database\Factories;
 
-use App\Book;
-use App\BookReview;
-use Faker\Generator as Faker;
+use App\Models\BookReview;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
 
-$factory->define(BookReview::class, function (Faker $faker) {
-    return [
-        'review' => $faker->randomElement(range(1, 10)),
-        'comment' => $faker->text(),
-    ];
-});
+/**
+ * @extends Factory<BookReview>
+ */
+class BookReviewFactory extends Factory
+{
+
+    public function definition(): array
+    {
+        return [
+            'review'  => $this->faker->randomElement(range(1, 10)),
+            'comment' => $this->faker->text(),
+        ];
+    }
+}
