@@ -11,20 +11,20 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        $users = factory(App\User::class, 5)->create();
-        $admin = factory(App\User::class, 1)->state('admin')->create();
+        $users = factory(\App\Models\User::class, 5)->create();
+        $admin = factory(\App\Models\User::class, 1)->admin()->create();
 
-        factory(App\Author::class, 15)->create()->each(function (App\Author $author) {
-            factory(App\Book::class, 3)->create()->each(function (App\Book $book) use ($author) {
+        factory(\App\Models\Author::class, 15)->create()->each(function (\App\Models\Author $author) {
+            factory(\App\Models\Book::class, 3)->create()->each(function (\App\Models\Book $book) use ($author) {
                 $book->authors()->saveMany([
                     $author,
                 ]);
             });
         });
 
-        \App\Book::all()->each(function (App\Book $book) use ($users) {
-            $reviews = factory(App\BookReview::class, 4)->make();
-            $reviews->each(function (\App\BookReview $review) use ($users) {
+        \App\Models\Book::all()->each(function (\App\Models\Book $book) use ($users) {
+            $reviews = factory(\App\Models\BookReview::class, 4)->make();
+            $reviews->each(function (\App\Models\BookReview $review) use ($users) {
                 $review->user()->associate($users->random());
             });
             $book->reviews()->saveMany($reviews);
